% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{authCookie}
\alias{authCookie}
\title{Gets that TT auth cookie, wrapper for the easy access to TT}
\usage{
authCookie()
}
\value{
TT
}
\description{
Gets that TT auth cookie, wrapper for the easy access to TT
}
