% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{executeSchedule}
\alias{executeSchedule}
\title{Execute schedule}
\usage{
executeSchedule(schedule.id, retry = FALSE)
}
\arguments{
\item{schedule.id}{The ID of the schedule you want to manage.}

\item{retry}{If true, applies the reschedule property if the schedule has it set. When not set, defaults to false.}
}
\description{
Executes GoodData schedule
}
\references{
GoodData API Documentation:
 https://help.gooddata.com/display/doc/API+Reference#/reference/scheduling-and-notifications/manage-schedule-executions/execute-a-schedule
}
