% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{getLastDefinition}
\alias{getLastDefinition}
\title{Gets last definition for a report object id}
\usage{
getLastDefinition(report.obj)
}
\arguments{
\item{report.obj}{integer id of the report object}
}
\value{
integer id of the latest report definitin object
}
\description{
Gets last definition for a report object id
}
