% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{getReportData}
\alias{getReportData}
\title{Gets report results as data.table}
\usage{
getReportData(uri)
}
\arguments{
\item{uri}{url for the report resutls}
}
\value{
data.table with report
}
\description{
Gets report results as data.table
}
