% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{getReportRawUri}
\alias{getReportRawUri}
\title{Function gets uri for export of raw report data}
\usage{
getReportRawUri(obj)
}
\arguments{
\item{obj}{integer id of the report definition, which refers to a version of a report}
}
\value{
uri of the report results that can be uploaded with the GET request
}
\description{
Function gets uri for export of raw report data
}
