% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{getResponseType}
\alias{getResponseType}
\title{Extracts repsonse type from the HTTP response}
\usage{
getResponseType(response)
}
\arguments{
\item{response}{http response object}
}
\value{
content type of the response
}
\description{
Extracts repsonse type from the HTTP response
}
