% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{processResponseError}
\alias{processResponseError}
\title{Stops the execution for response with bad status codes}
\usage{
processResponseError(response)
}
\description{
Stops the execution for response with bad status codes
}
