% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{superSecuredToken}
\alias{superSecuredToken}
\title{Gets the  Super Secured Token (SST) for Login}
\usage{
superSecuredToken()
}
\value{
STT
}
\description{
Gets the  Super Secured Token (SST) for Login
}
\keyword{internal}
