% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{temporaryToken}
\alias{temporaryToken}
\title{Gets the TT (Temporary Token) for a given SST}
\usage{
temporaryToken(sst)
}
\arguments{
\item{sst}{SST}
}
\value{
TT
}
\description{
Gets the TT (Temporary Token) for a given SST
}
