% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{authCookie}
\alias{authCookie}
\title{Gets that Temporary Token (TT) auth cookie, wrapper for the easy access to TT}
\usage{
authCookie()
}
\value{
Temporary Token (TT)
}
\description{
Gets that Temporary Token (TT) auth cookie, wrapper for the easy access to TT
}
\keyword{internal}
