% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{processResponse}
\alias{processResponse}
\title{Processes http response and returns a list of results}
\usage{
processResponse(response)
}
\arguments{
\item{response}{http response from the GoodData API}
}
\value{
list of resutls parsed from the response
}
\description{
Processes http response and returns a list of results
}
\keyword{internal}
