% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{processResponseError}
\alias{processResponseError}
\title{Stops the execution for response with bad status codes}
\usage{
processResponseError(response)
}
\arguments{
\item{response}{HTTP response that has error status code}
}
\description{
Stops the execution for response with bad status codes
}
\keyword{internal}
