% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{temporaryToken}
\alias{temporaryToken}
\title{Gets the Temporary Token (TT) for a given Super Secured Token (SST)}
\usage{
temporaryToken(sst)
}
\arguments{
\item{sst}{SST}
}
\value{
TT
}
\description{
Gets the Temporary Token (TT) for a given Super Secured Token (SST)
}
\keyword{internal}
